Add files via upload
I've stripped out all identifying marks as I've used this at my place of work. But it's a basic tool that was fun to make. The PyQt5 part gives a simple tool to users, then Powershell, at a preset time via a job, picks up the attendance and e-mails it out.